Fix typescript errors
When fixing typescript errors I also ran prettier over the files and
removed unused imports. The Checkbox size prop had no impact on styling
and was causing a typescript error. So I removed it and references to
a size prop on where it was being used. FormErrorIcon is an old
Chakra UI component from version 1 and doesn't exist in the lastest
version so I have removed it.