feat: add container to design system
Currently there exists a grid component which is not responsive.
It requires entry of column number, which is difficult to make
responsive.

Using a flex box means the page will be responsive and match the designs.
While number of columns cannot be specified, there is currently no use
of more than 3 columns in the designs. Use of a container may therefore
have more pros than cons.